test: consolidated all conftests into a single one.
